% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_protected_areas.R
\name{addNewProtectedArea}
\alias{addNewProtectedArea}
\title{addNewProtectedArea applies consistent leaflet behaviour to each Expand button}
\usage{
addNewProtectedArea(map, data, layerId, layerList)
}
\description{
addNewProtectedArea applies consistent leaflet behaviour to each Expand button
}
