% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bowen_map_toppct}
\alias{bowen_map_toppct}
\title{Bowen Map Plotting for Top Percentage}
\usage{
bowen_map_toppct(
  raster_layer,
  title,
  subtitle,
  caption,
  legend_label,
  pal = "ag_GrnYl"
)
}
\arguments{
\item{raster_layer}{SpatRaster, input raster}

\item{title}{Text, title for map}

\item{subtitle}{Text, subtitle for map}

\item{caption}{Text, caption for map}

\item{legend_label}{Text, legend label}

\item{pal}{Palette, from colorspace package}
}
\value{
ggplot
}
\description{
Bowen Map Plotting for Top Percentage
}
