% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sdm.R
\name{bowen_rasterize}
\alias{bowen_rasterize}
\title{Rasterize to Bowen Island Mask}
\usage{
bowen_rasterize(sf)
}
\arguments{
\item{sf}{input sf object to rasterize}
}
\description{
Rasterize to Bowen Island Mask
}
