% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sdm.R
\name{bowen_sf_to_rast}
\alias{bowen_sf_to_rast}
\title{Convert sf polygons to terra rast objects}
\usage{
bowen_sf_to_rast(input_sf, varname = "Present")
}
\arguments{
\item{input_sf}{sf object to rasterize to Bowen Island extent}

\item{varname}{varname to describe rast band}
}
\value{
terra rast object
}
\description{
Convert sf polygons to terra rast objects
}
