% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_elements.R
\name{docs_link}
\alias{docs_link}
\title{docs_link}
\usage{
docs_link(href = "https://palen-lab.github.io/bowen.biodiversity.webapp/")
}
\arguments{
\item{href}{string, link to documentation page}
}
\value{
UI component
}
\description{
docs_link
}
