% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hatched.sf}
\alias{hatched.sf}
\title{Hatch sf polygon}
\usage{
hatched.sf(poly_sf, density = 1, pattern = "right2left")
}
\arguments{
\item{poly_sf}{Polygon or MultiPolygon sf object}

\item{density}{Density of lines}

\item{pattern}{Line pattern}
}
\value{
sf Multilinestring sf object
}
\description{
Hatch sf polygon
}
