% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{higher_res_edges}
\alias{higher_res_edges}
\title{Deal with edges}
\usage{
higher_res_edges(raster_layer)
}
\arguments{
\item{raster_layer}{rast}
}
\value{
rast
}
\description{
Increases resolution for the purposes of plotting only. Disaggregates the
input raster without manipulating values. Raster is then masked by the Bowen
Island ocean sf, so that the jagged appearance along coastline is less
pronounced in produced maps.
}
