% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_protected_areas.R
\name{highlightProtectedArea}
\alias{highlightProtectedArea}
\title{Update Leaflet Map for each Existing Protected Area ####}
\usage{
highlightProtectedArea(map, data)
}
\description{
Update Leaflet Map for each Existing Protected Area ####
}
