% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sdm.R
\name{normalize}
\alias{normalize}
\title{Normalize on 0 to 1 scale}
\usage{
normalize(input_rast)
}
\arguments{
\item{input_rast}{input raster}
}
\value{
terra rast where the values are normalized
}
\description{
Normalize on 0 to 1 scale
}
