% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sdm.R
\name{raster_overlap_ratio}
\alias{raster_overlap_ratio}
\title{Provide ratio of overlap between two rasters (in terms of NA vs nonNA)}
\usage{
raster_overlap_ratio(top_rast, area_rast)
}
\description{
Provide ratio of overlap between two rasters (in terms of NA vs nonNA)
}
