% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sdm.R
\name{raster_prep_bowen}
\alias{raster_prep_bowen}
\title{Prepare rasters to match Bowen Island}
\usage{
raster_prep_bowen(input_rast, bowen_boundary, bowen_shoreline, crs)
}
\arguments{
\item{input_rast, }{SpatRaster}
}
\value{
SpatRaster
}
\description{
Prepare rasters to match Bowen Island
}
