% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sdm.R
\name{remove_by_quantile}
\alias{remove_by_quantile}
\title{Only keep top or bottom cell values by quantile}
\usage{
remove_by_quantile(x, prob, direction = "top")
}
\arguments{
\item{x}{SpatRaster}

\item{prob}{Quantile between 0 and 1}

\item{direction}{"top" or "bottom", for direction of keep / discard cell values}
}
\value{
SpatRaster
}
\description{
Only keep top or bottom cell values by quantile
}
