% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sdm.R
\name{sdm_to_species_richness}
\alias{sdm_to_species_richness}
\title{Create species richness map from SDM rasters}
\usage{
sdm_to_species_richness(SDM_stack, presence_threshold = 0.7)
}
\arguments{
\item{sdm_stack}{raster stack, to be combined into a richness map}

\item{threshold}{numeric, value from 0 to 1 corresponding to probability
threshold in the cells of the SDM to include into species richness map}
}
\value{
SpatRaster, with species present in each cell
}
\description{
Create species richness map from SDM rasters
}
