% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overlay.R
\name{top_pct_zonation_vect}
\alias{top_pct_zonation_vect}
\title{top_pct_zonation_vect}
\usage{
top_pct_zonation_vect(map, data, top_pct)
}
\description{
For providing the greyed out effect on areas not in top %
}
