% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overlay.R
\name{values_in_out}
\alias{values_in_out}
\title{Overlap}
\usage{
values_in_out(area_mask, selected_mask, input_rast)
}
\arguments{
\item{area_mask}{SpatRaster, has values for study area and NA for
outside study area}

\item{selected_mask}{SpatRaster, has values for selected pixels
within study area, and NA for not selected pixels and outside study
area}

\item{input_rast}{SpatRaster, has values that will be extracted for
inside and outside study area.}
}
\value{
named list, for values inside and outside selected pixels
from input_rast. Removes NA values.
}
\description{
Overlap
}
